+++ src/events/SunshineEvent.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <Geode/loader/Event.hpp>
+
+
+struct SunshineEvent final : geode::Event {
+	SunshineEvent() : geode::Event() {}
+};