+++ src/hooks/MenuLayer.cpp
@@ -1,0 +1,34 @@
+#include <Geode/modify/MenuLayer.hpp>
+
+#include "SunshineEvent.hpp"
+
+using namespace geode::prelude;
+
+
+class $modify(MenuLayer) {
+	$override
+	bool init() {
+		if (!MenuLayer::init())
+			return false;
+
+		if (!Mod::get()->getSettingValue<bool>("general-main-menu"))
+			return true;
+
+		auto lebrutton = CCMenuItemExt::createSpriteExtra(
+			CircleButtonSprite::createWithSpriteFrameName(
+				"sunshine-small.png"_spr,
+				1.f
+			),
+			[](CCMenuItemSpriteExtra*) {
+				SunshineEvent().post();
+			}
+		);
+		lebrutton->setID("lebrutton"_spr);
+
+		auto menu = this->getChildByID("bottom-menu");
+		menu->addChild(lebrutton);
+		menu->updateLayout();
+
+		return true;
+	}
+};