+++ changelog.md
@@ -1,2 +1,3 @@
 # 1.0.0
 - Initial release
+- In the future I'll add optional Custom ~~Keybinds~~ Keyrings support (trust), it's just that lead devs don't want to release optional API for it